Modularize code a bit
